requirements.txt
keras==2.7.0
numpy==1.20.3
pandas==1.3.5
Pillow==9.0.1
streamlit==1.5.0
mysql-connector-python==8.0.26
matplotlib==3.4.2
mysql-connector-python==8.0.26
scikit-learn==1.0
seaborn==0.11.2
pickle-mixin==1.0.2
tensorflow-cpu==2.6.0